Update (2024.10.16)
34617: LA port of 8333354: ubsan: frame.inline.hpp:91:25: and src/hotspot/share/runtime/frame.inline.hpp:88:29: runtime error: member call on null pointer of type 'const struct SmallRegisterMap'
34182: LA port of 8329258: TailCall should not use frame pointer register for jump target
34100: LA port of 8325469: Freeze/Thaw code can crash in the presence of OSR frames
31470: LA port of 8309685: Fix -Wconversion warnings in assembler and register code
31471: LA port of 8310906: Fix -Wconversion warnings in runtime, oops and some code header files.

+++ src/hotspot/cpu/loongarch/register_loongarch.hpp
@@ -55,7 +55,7 @@ class Register {
 
    public:
     // accessors
-    constexpr int raw_encoding() const { return this - first(); }
+    constexpr int raw_encoding() const { return checked_cast<int>(this - first()); }
     constexpr int     encoding() const { assert(is_valid(), "invalid register"); return raw_encoding(); }
     constexpr bool    is_valid() const { return 0 <= raw_encoding() && raw_encoding() < number_of_registers; }
 
@@ -225,7 +225,7 @@ class FloatRegister {
 
    public:
     // accessors
-    constexpr int raw_encoding() const { return this - first(); }
+    constexpr int raw_encoding() const { return checked_cast<int>(this - first()); }
     constexpr int     encoding() const { assert(is_valid(), "invalid register"); return raw_encoding(); }
     constexpr bool    is_valid() const { return 0 <= raw_encoding() && raw_encoding() < number_of_registers; }
 
@@ -386,7 +386,7 @@ class ConditionalFlagRegister {
 
    public:
     // accessors
-    int raw_encoding() const { return this - first(); }
+    int raw_encoding() const { return checked_cast<int>(this - first()); }
     int encoding() const     { assert(is_valid(), "invalid register"); return raw_encoding(); }
     bool is_valid() const    { return 0 <= raw_encoding() && raw_encoding() < number_of_registers; }
